2037: separate the TCAP functionality from the TCA (tca.go -> tcap.go)
